Update your local repository in CONTRIBUTING.md
Enrich the contect of CONTRIBUTING.md with a new section called 'Getting up-to-date'.
Address an important issue for a large collaborrative team, which is the frequent
updates of local repositories of the collaborators.

Taking into consideration, the current working branch of the developper, different commands
were provided on how to update his/her branch and acquire the latest version of the project.

Important git commands included to this section:

- rebase

- merge